+++ README.md
@@ -1,1 +1,3 @@
 # Online_Chat_Messenger
+
+## 議事録 : https://docs.google.com/document/d/1N-IGM2l2ffOHJxVztnwcg26gQbwUOPm8h2RvozZYxo8/edit?usp=sharing
